Updated docstrings and combined all video streaming classes into one file.
